fix: enforce benchmark and opcode gates
