+++ .github/workflows/ci.yml
@@ -66,7 +66,6 @@ jobs:
 
   benchmark:
     needs: check
-    continue-on-error: true
     name: Benchmark Report
     runs-on: ubuntu-24.04
     steps:
