--- benchmarks/compare_disabled_test.go
@@ -1,7 +1,0 @@
-//go:build !compare
-
-package benchmarks
-
-import "testing"
-
-func benchmarkCompare(*testing.B, benchmarkComparison, int32) {}
